+++ packages/asc-ui/src/components/Menu/MenuInline/MenuInline.tsx
@@ -1,18 +1,20 @@
 import React from 'react'
 import MenuInlineStyle, { Props } from './MenuInlineStyle'
 import MenuContext from '../MenuContext'
-import BackDrop from '../../BackDrop/BackDrop'
+import BackDrop, { Props as BackDropProps } from '../../BackDrop/BackDrop'
 import useDebounce from '../../../utils/hooks/useDebounce'
 
-const MenuInline: React.FC<Props> = ({
+const MenuInline: React.FC<Props & BackDropProps> = ({
   children,
   onExpand,
   hasBackDrop,
+  backdropOpacity,
+  zIndexOffset,
   ...otherProps
 }) => {
   const [menuOpen, setMenuOpen] = React.useState(false)
 
-  // prevents the backdrop from flickering when you hover to the next menu items
+  // prevents the component from flickering when you hover to the next menu items
   const handleOnExpand = useDebounce((expand: boolean) => {
     setMenuOpen(expand)
 
@@ -29,7 +31,11 @@ const MenuInline: React.FC<Props> = ({
         {children}
       </MenuInlineStyle>
       {hasBackDrop && menuOpen && (
-        <BackDrop onClick={handleOnExpand} hideOverFlow={false} />
+        <BackDrop
+          onClick={handleOnExpand}
+          backdropOpacity={backdropOpacity}
+          zIndexOffset={zIndexOffset}
+        />
       )}
     </MenuContext.Provider>
   )