+++ packages/asc-ui/src/components/Toggle/Toggle.tsx
@@ -6,7 +6,7 @@ import useActionOnEscape from '../../utils/hooks/useActionOnEscape'
 import ToggleButton, {
   Props as ToggleButtonProps,
 } from '../Button/ToggleButton/ToggleButton'
-import BackDrop from '../BackDrop/BackDrop'
+import BackDrop, { Props as BackDropProps } from '../BackDrop/BackDrop'
 
 export type ToggleHandlerProps = {
   as?: keyof JSX.IntrinsicElements | React.ComponentType<any>
@@ -23,7 +23,8 @@ export type Props = {
   rotateOnOpen?: number
   as?: keyof JSX.IntrinsicElements | React.ComponentType<any>
 } & ToggleStyleProps &
-  ToggleHandlerProps
+  ToggleHandlerProps &
+  BackDropProps
 
 // Todo: refactor this to Collapse component https://github.com/Amsterdam/amsterdam-styled-components/issues/379
 const Toggle: React.FC<Props & HTMLAttributes<HTMLDivElement>> = ({
@@ -39,6 +40,8 @@ const Toggle: React.FC<Props & HTMLAttributes<HTMLDivElement>> = ({
   ToggleHandler,
   title,
   hasBackDrop,
+  backdropOpacity,
+  zIndexOffset,
   ariaLabel,
   ...otherProps
 }) => {
@@ -128,7 +131,11 @@ const Toggle: React.FC<Props & HTMLAttributes<HTMLDivElement>> = ({
       )}
       {render ? children : conditionalRenderedChildren}
       {hasBackDrop && open && (
-        <BackDrop onClick={handleOnClick} hideOverFlow={false} />
+        <BackDrop
+          onClick={handleOnClick}
+          zIndexOffset={zIndexOffset}
+          backdropOpacity={backdropOpacity}
+        />
       )}
     </ToggleStyle>
   )