+++ packages/asc-ui/src/utils/deprecatedWarning.ts
@@ -1,0 +1,8 @@
+/**
+ * A general deprecated console warning avoided by lint
+ */
+
+export default (message: String) => {
+  // eslint-disable-next-line
+  console.warn(message)
+}